Fix secp256k1 build failure
The pod that we were using is excluding all files that have "test" in the name. This is causing a build issue because secp256k1.c has a reference to selftest.h.

Fixed it by using a diferent cocoapod that also points to bitcoin-core/secp256k1 that is not filtering our that file.